refactor: extract types from QueryFilterModel to separate files
