+++ ui/apps/vue-mri-ui-lib/src/query-filter/types/AtlasTypes.ts
@@ -1,0 +1,9 @@
+/**
+ * Atlas cohort definition types for OHDSI Atlas integration
+ * 
+ * NOTE: NumericRange and DemographicCriteria are defined in AtlasCohortDefinition.ts
+ * This file is reserved for additional Atlas-specific types that may be needed.
+ */
+
+// Re-export types from AtlasCohortDefinition for convenience
+export type { NumericRange, DemographicCriteria } from '../models/AtlasCohortDefinition'
