+++ ui/apps/vue-mri-ui-lib/src/query-filter/types/ConceptSetTypes.ts
@@ -2,7 +2,7 @@
  * TypeScript interfaces for concept set functionality
  */
 
-export interface ConceptSetItem {
+export interface ConceptSetItemDisplay {
   value: string
   text?: string
   display_value?: string
@@ -46,7 +46,7 @@ export interface ConceptSetConfig {
 }
 
 export interface ConceptSetDomainValues {
-  values: ConceptSetItem[]
+  values: ConceptSetItemDisplay[]
   isLoading: boolean
   loadedStatus: 'NO_RESULTS' | 'HAS_RESULTS' | 'TOO_MANY_RESULTS'
 }
@@ -55,7 +55,7 @@ export interface TagInputModel {
   id: string
   props: {
     type: string
-    value: ConceptSetItem[]
+    value: ConceptSetItemDisplay[]
     attributePath: string
     domainFilter: string
     standardConceptCodeFilter: string
@@ -63,7 +63,7 @@ export interface TagInputModel {
 }
 
 export interface ConceptSetAction {
-  values?: ConceptSetItem
+  values?: ConceptSetItemDisplay
   config?: ConceptSetConfig
   componentType?: string // Type to determine modal mode (concept vs conceptSet)
   attributeId?: string // ID of the attribute that triggered this action
@@ -85,4 +85,65 @@ export interface CreateConceptSetRequest {
   name: string
   shared: boolean
   userName: string
+}
+
+// Types moved from QueryFilterModel.ts
+export interface StoredConceptItem {
+  value: string
+  text: string
+  display_value: string
+  conceptId: number
+  domainId?: string | undefined
+  system?: string | undefined
+  conceptClassId?: string | undefined
+  standardConcept?: string | undefined
+  concept?: string | undefined
+  code?: string | undefined
+  validStartDate?: string | undefined
+  validEndDate?: string | undefined
+  validity?: string | undefined
+  useDescendants?: boolean | undefined
+  useMapped?: boolean | undefined
+  isExcluded?: boolean | undefined
+  score?: number | undefined
+  conceptName?: string | undefined
+}
+
+export interface ConceptSetDetailConcept {
+  CONCEPT_ID: number
+  CONCEPT_NAME: string
+  STANDARD_CONCEPT: string
+  STANDARD_CONCEPT_CAPTION: string
+  INVALID_REASON: string
+  INVALID_REASON_CAPTION: string
+  CONCEPT_CODE: string
+  DOMAIN_ID: string
+  VOCABULARY_ID: string
+  CONCEPT_CLASS_ID: string
+}
+
+export interface ConceptSetDetail {
+  concept: ConceptSetDetailConcept
+  isExcluded: boolean
+  includeDescendants: boolean
+  includeMapped: boolean
+}
+
+export interface SelectedConceptSetConcept {
+  id: number
+  useMapped: boolean
+  isExcluded: boolean
+  useDescendants: boolean
+}
+
+export interface SelectedConceptSet {
+  value: number
+  text: string
+  display_value: string
+  conceptIds: number[]
+  concepts: SelectedConceptSetConcept[]
+  shared: boolean
+  userName: string
+  createdDate: string
+  modifiedDate: string
 }
